requirements.txt
langgraph==0.5.2
langchain==0.3.26
langchain-openai==0.3.27
langchain-community==0.3.27
langchain-core==0.3.68
chromadb==0.5.0
sentence-transformers==2.7.0
python-dotenv==1.0.0
openai==1.51.0
pydantic>=2.7.4
fastapi==0.111.0
uvicorn==0.30.0 